feat: Improve tlcsdm widgets
