+++ bundles/com.tlcsdm.tlstudio.widgets/src/com/tlcsdm/tlstudio/widgets/color/SaturationBrightness.java
@@ -7,11 +7,7 @@ import org.eclipse.swt.graphics.ImageData;
 import org.eclipse.swt.graphics.PaletteData;
 import org.eclipse.swt.graphics.RGB;
 import org.eclipse.swt.graphics.Rectangle;
-import org.eclipse.swt.layout.GridData;
-import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.Display;
-import org.eclipse.swt.widgets.Shell;
 
 import com.tlcsdm.tlstudio.widgets.custom.AbstractCustomCanvas;
 
@@ -57,27 +53,4 @@ public class SaturationBrightness extends AbstractCustomCanvas {
 		saturationImage.dispose();
 	}
 
-	public static void main(String[] args) {
-		Display display = new Display();
-
-		Shell shell = new Shell(display);
-		shell.setText("Hue");
-		GridLayout gridLayout = new GridLayout(2, false);
-		gridLayout.verticalSpacing = 10;
-		gridLayout.marginWidth = gridLayout.marginHeight = 16;
-		shell.setLayout(gridLayout);
-
-		SaturationBrightness cd = new SaturationBrightness(shell, SWT.NONE);
-		GridData data = new GridData(SWT.RIGHT, SWT.CENTER, false, false, 2, 1);
-		data.widthHint = 500;
-		data.heightHint = 500;
-		cd.setLayoutData(data);
-
-		shell.open();
-		while (!shell.isDisposed()) {
-			if (!display.readAndDispatch())
-				display.sleep();
-		}
-		display.dispose();
-	}
 }
