+++ examples/com.tlcsdm.tlstudio.widgets.snippets/src/com/tlcsdm/tlstudio/widgets/example/color/SaturationBrightnessSnippet.java
@@ -1,0 +1,36 @@
+package com.tlcsdm.tlstudio.widgets.example.color;
+
+import org.eclipse.swt.SWT;
+import org.eclipse.swt.layout.GridData;
+import org.eclipse.swt.layout.GridLayout;
+import org.eclipse.swt.widgets.Display;
+import org.eclipse.swt.widgets.Shell;
+
+import com.tlcsdm.tlstudio.widgets.color.SaturationBrightness;
+
+public class SaturationBrightnessSnippet {
+
+	public static void main(String[] args) {
+		Display display = new Display();
+
+		Shell shell = new Shell(display);
+		shell.setText("Hue");
+		GridLayout gridLayout = new GridLayout(2, false);
+		gridLayout.verticalSpacing = 10;
+		gridLayout.marginWidth = gridLayout.marginHeight = 16;
+		shell.setLayout(gridLayout);
+
+		SaturationBrightness cd = new SaturationBrightness(shell, SWT.NONE);
+		GridData data = new GridData(SWT.RIGHT, SWT.CENTER, false, false, 2, 1);
+		data.widthHint = 500;
+		data.heightHint = 500;
+		cd.setLayoutData(data);
+
+		shell.open();
+		while (!shell.isDisposed()) {
+			if (!display.readAndDispatch())
+				display.sleep();
+		}
+		display.dispose();
+	}
+}
