+++ examples/com.tlcsdm.tlstudio.widgets.snippets/src/com/tlcsdm/tlstudio/widgets/example/custom/CollapsibleCompositeSnippet.java
